requirements.txt
Django==4.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
pytest-django==4.5.2
pytest==7.4.0
Pillow==10.0.0